feat(racingcar): CarName 클래스 생성
