+++ src/main/java/racingcar/Car.java
@@ -1,7 +1,5 @@
 package racingcar;
 
-import org.springframework.util.StringUtils;
-
 public class Car {
     private final CarName name;
     private int position;
@@ -25,25 +23,4 @@ public class Car {
         return name.getValue();
     }
 
-    class CarName {
-        private final String name;
-
-        public CarName(String name) {
-            validate(name);
-            this.name = name;
-        }
-
-        private void validate(String name) {
-            if (!StringUtils.hasText(name)) {
-                throw new IllegalArgumentException();
-            }
-            if (name.length() > 5) {
-                throw new IllegalArgumentException();
-            }
-        }
-
-        public String getValue() {
-            return name;
-        }
-    }
 }
